+++ AvazehApiClient/DataAccess/CollectionManagers/InvoiceCollectionManagerAsync.cs
@@ -144,14 +144,14 @@ public class InvoiceCollectionManagerAsync : IInvoiceCollectionManager
         return collection is null ? null : collection;
     }
 
-    public async Task<double> GetCustomerTotalBalanceById(int CustomerId, int InvoiceId = 0)
+    public async Task<decimal> GetCustomerTotalBalanceById(int CustomerId, int InvoiceId = 0)
     {
-        return (double)await ApiProcessor.GetValueOrNullAsync<double>(Key + "/CustomerBalance", CustomerId, InvoiceId);
+        return (decimal)await ApiProcessor.GetValueOrNullAsync<decimal>(Key + "/CustomerBalance", CustomerId, InvoiceId);
     }
 
-    public async Task<double> GetInvoicePrevTotalBalanceById(int InvoiceId) //Calculates balance for prev id of this invoice.
+    public async Task<decimal> GetInvoicePrevTotalBalanceById(int InvoiceId) //Calculates balance for prev id of this invoice.
     {
-        return (double)await ApiProcessor.GetValueOrZeroAsync<double>(Key + "/PrevBalance", InvoiceId);
+        return (decimal)await ApiProcessor.GetValueOrZeroAsync<decimal>(Key + "/PrevBalance", InvoiceId);
     }
 
     public async Task<List<UserDescriptionModel>> GetUserDescriptions()
