+++ SharedLibrary/DalModels/InvoiceItemForPrintModel.cs
@@ -6,9 +6,9 @@ public class InvoiceItemForPrintModel
 {
     public int Id { get; set; }
     public string ProductName { get; set; }
-    public double SellPrice { get; set; }
+    public long SellPrice { get; set; }
     public string CountString { get; set; }
-    public double TotalPrice { get; set; }
+    public decimal TotalPrice { get; set; }
     public string DateCreated { get; set; }
     public string TimeCreated { get; set; }
     public string DateUpdated { get; set; }
